split code from compactor
