+++ src/storage/src/hummock/shared_buffer/shared_buffer_uploader.rs
@@ -19,8 +19,9 @@ use risingwave_hummock_sdk::filter_key_extractor::FilterKeyExtractorManagerRef;
 use risingwave_hummock_sdk::{HummockEpoch, LocalSstableInfo};
 use risingwave_rpc_client::HummockMetaClient;
 
-use crate::hummock::compaction_executor::CompactionExecutor;
-use crate::hummock::compactor::{Compactor, CompactorContext};
+use crate::hummock::compactor::{
+    compact_shared_buffer_by_compaction_group, CompactionExecutor, CompactorContext,
+};
 use crate::hummock::conflict_detector::ConflictDetector;
 use crate::hummock::shared_buffer::OrderSortedUncommittedData;
 use crate::hummock::{HummockResult, MemoryLimiter, SstableIdManagerRef, SstableStoreRef};
@@ -37,8 +38,7 @@ pub struct SharedBufferUploader {
     hummock_meta_client: Arc<dyn HummockMetaClient>,
     stats: Arc<StateStoreMetrics>,
     compaction_executor: Option<Arc<CompactionExecutor>>,
-    local_object_store_compactor_context: Arc<CompactorContext>,
-    remote_object_store_compactor_context: Arc<CompactorContext>,
+    compactor_context: Arc<CompactorContext>,
 }
 
 impl SharedBufferUploader {
@@ -60,18 +60,7 @@ impl SharedBufferUploader {
         };
         // not limit memory for uploader
         let memory_limiter = Arc::new(MemoryLimiter::new(u64::MAX - 1));
-        let local_object_store_compactor_context = Arc::new(CompactorContext {
-            options: options.clone(),
-            hummock_meta_client: hummock_meta_client.clone(),
-            sstable_store: sstable_store.clone(),
-            stats: stats.clone(),
-            is_share_buffer_compact: true,
-            compaction_executor: compaction_executor.as_ref().cloned(),
-            filter_key_extractor_manager: filter_key_extractor_manager.clone(),
-            memory_limiter: memory_limiter.clone(),
-            sstable_id_manager: sstable_id_manager.clone(),
-        });
-        let remote_object_store_compactor_context = Arc::new(CompactorContext {
+        let compactor_context = Arc::new(CompactorContext {
             options: options.clone(),
             hummock_meta_client: hummock_meta_client.clone(),
             sstable_store: sstable_store.clone(),
@@ -89,8 +78,7 @@ impl SharedBufferUploader {
             hummock_meta_client,
             stats,
             compaction_executor,
-            local_object_store_compactor_context,
-            remote_object_store_compactor_context,
+            compactor_context,
         }
     }
 }
@@ -99,7 +87,6 @@ impl SharedBufferUploader {
     pub async fn flush(
         &self,
         epoch: HummockEpoch,
-        is_local: bool,
         payload: UploadTaskPayload,
     ) -> HummockResult<Vec<LocalSstableInfo>> {
         if payload.is_empty() {
@@ -107,11 +94,7 @@ impl SharedBufferUploader {
         }
 
         // Compact buffers into SSTs
-        let mem_compactor_ctx = if is_local {
-            self.local_object_store_compactor_context.clone()
-        } else {
-            self.remote_object_store_compactor_context.clone()
-        };
+        let mem_compactor_ctx = self.compactor_context.clone();
 
         // Set a watermark SST id for this epoch to prevent full GC from accidentally deleting SSTs
         // for in-progress write op. The watermark is invalidated when the epoch is
@@ -121,9 +104,7 @@ impl SharedBufferUploader {
             .add_watermark_sst_id(Some(epoch))
             .await?;
 
-        let tables =
-            Compactor::compact_shared_buffer_by_compaction_group(mem_compactor_ctx, payload)
-                .await?;
+        let tables = compact_shared_buffer_by_compaction_group(mem_compactor_ctx, payload).await?;
 
         let uploaded_sst_info = tables.into_iter().collect();
